making right structure
