+++ parentheses/parentheses.go
@@ -1,10 +1,8 @@
 package parentheses
 
 import (
-	"fmt"
 	"math/rand"
-	"net/http"
-	"strconv"
+	"strings"
 )
 
 func IsBalanced(s string) bool {
@@ -32,36 +30,22 @@ func IsBalanced(s string) bool {
 }
 
 func GenerateSequence(length int) string {
-	var parentheses string
+	var builder strings.Builder
 	for i := 0; i < length; i++ {
 		switch rand.Perm(6)[0] {
 		case 0:
-			parentheses += "("
+			builder.WriteRune('(')
 		case 1:
-			parentheses += ")"
+			builder.WriteRune(')')
 		case 2:
-			parentheses += "["
+			builder.WriteRune('[')
 		case 3:
-			parentheses += "]"
+			builder.WriteRune(']')
 		case 4:
-			parentheses += "{"
+			builder.WriteRune('{')
 		case 5:
-			parentheses += "}"
+			builder.WriteRune('}')
 		}
 	}
-	return parentheses
-}
-
-func GenerateSequenceHandler(w http.ResponseWriter, r *http.Request) {
-	lengthParam := r.URL.Query().Get("n")
-
-	length, err := strconv.Atoi(lengthParam)
-	if err != nil {
-		http.Error(w, "Invalid length parameter", http.StatusBadRequest)
-		return
-	}
-
-	parentheses := GenerateSequence(length)
-
-	fmt.Fprint(w, parentheses)
+	return builder.String()
 }
